Fixed user context issue
