+++ src/views/dashboard/sections/HandleGroupView.tsx
@@ -1,5 +1,5 @@
 import React, { useContext } from "react";
-import { UserContext } from "GlobalContext/UserContext";
+import { TestUserContext } from "GlobalContext/TestUserContext";
 import { DashboardUser } from "./dashboard-user/DashboardUser";
 import { DashboardAdmin } from "./dashboard-admin/DashboardAdmin";
 
@@ -7,7 +7,7 @@ interface Props {
   match: any;
 }
 const HandleGroupView: React.FC<Props> = props => {
-  const state: { userId: number; groups: any } = useContext(UserContext);
+  const state: { userId: number; groups: any } = useContext(TestUserContext);
   const group = state.groups.find(
     (group: { groupId: any }) =>
       group.groupId.toString() === props.match.params.groupId,
